chore: type work + move schema types to common
